chore: refactoring partion interface and centralizing in internal folder
