+++ datastreams/datastream.go
@@ -2,6 +2,7 @@ package datastreams
 
 import (
 	"context"
+	"github.com/elastiflow/pipelines/datastreams/internal/pipes"
 	"sync"
 )
 
@@ -190,7 +191,7 @@ func (p DataStream[T]) FanOut(
 	param := applyParams(params...)
 	nextPipe, outChannels := p.nextT(fanOut, param)
 	p.incrementWaitGroup(1)
-	go func(inStream <-chan T, outStreams senders[T]) {
+	go func(inStream <-chan T, outStreams pipes.Senders[T]) {
 		if p.wg != nil {
 			defer p.wg.Done()
 		}
@@ -290,7 +291,7 @@ func (p DataStream[T]) Broadcast(
 	param := applyParams(params...)
 	nextPipe, outChannels := p.nextT(broadcast, param)
 	p.incrementWaitGroup(1)
-	go func(inStream <-chan T, outStreams senders[T]) {
+	go func(inStream <-chan T, outStreams pipes.Senders[T]) {
 		if p.wg != nil {
 			defer p.wg.Done()
 		}
@@ -342,7 +343,7 @@ func (p DataStream[T]) Tee(
 	return nextPipe1, nextPipe2
 }
 
-func (p DataStream[T]) nextT(pipeType pipeType, params Params) (DataStream[T], pipes[T]) {
+func (p DataStream[T]) nextT(pipeType pipeType, params Params) (DataStream[T], pipes.Pipes[T]) {
 	return next[T](pipeType, params, len(p.inStreams), p.ctx, p.errStream, p.wg)
 }
 
@@ -438,7 +439,7 @@ func next[T any](
 	ctx context.Context,
 	errStream chan<- error,
 	wg *sync.WaitGroup,
-) (DataStream[T], pipes[T]) {
+) (DataStream[T], pipes.Pipes[T]) {
 	switch pipeType {
 	case fanOut, broadcast:
 		chanCount = params.Num
@@ -446,7 +447,7 @@ func next[T any](
 		chanCount = 1
 	default:
 	}
-	streams := make(pipes[T], chanCount)
+	streams := make(pipes.Pipes[T], chanCount)
 	streams.Initialize(params.BufferSize)
 	return DataStream[T]{
 			ctx:       ctx,
