+++ datastreams/internal/partition/base_test.go
@@ -1,0 +1,1 @@
+package partition
