+++ datastreams/windower/interval.go
@@ -3,6 +3,7 @@ package windower
 import (
 	"context"
 	"github.com/elastiflow/pipelines/datastreams/internal/partition"
+	"github.com/elastiflow/pipelines/datastreams/internal/pipes"
 	"time"
 )
 
@@ -17,37 +18,19 @@ type timedInterval[T any, R any] struct {
 // arrive in a window, we publish an empty batch.
 func NewInterval[T any, R any](
 	ctx context.Context,
+	outChannels pipes.Pipes[R],
 	procFunc func([]T) (R, error),
 	errs chan<- error,
 	interval time.Duration,
 ) partition.Partition[T, R] {
 	w := &timedInterval[T, R]{
-		Base:     partition.NewBase[T, R](ctx, errs),
+		Base:     partition.NewBase[T, R](ctx, outChannels, errs),
 		interval: interval,
 		procFunc: procFunc,
 	}
 	// Start the background ticker right away
 	go w.startInterval()
 	return w
-}
-
-// NewIntervalFactory constructs a factory function that can be used
-// to create new instances of timedInterval. This is useful for
-// creating multiple instances with the same processing function and
-// interval duration.
-func NewIntervalFactory[T any, R any](
-	procFunc func([]T) (R, error),
-	interval time.Duration,
-) partition.Factory[T, R] {
-	return func(ctx context.Context, errs chan<- error) partition.Partition[T, R] {
-		return NewInterval[T, R](ctx, procFunc, errs, interval)
-	}
-}
-
-// Push just appends to the in-memory buffer.
-// The separate ticker goroutine will publish periodically.
-func (t *timedInterval[T, R]) Push(item T, _ time.Time) {
-	t.Batch.Push(item)
 }
 
 // startInterval continuously ticks every `w.interval`, publishes whatever
@@ -61,9 +44,26 @@ func (t *timedInterval[T, R]) startInterval() {
 			if t.Batch.Len() == 0 {
 				continue
 			}
-			go partition.Flush[[]T, R](t.Ctx, t.Batch.Next(), t.Out, t.procFunc, t.Errs)
+			go t.Flush(t.Ctx, t.Batch.Next(), t.procFunc, t.Errs)
 		case <-t.Ctx.Done():
 			return
 		}
+	}
+}
+
+// NewIntervalFactory constructs a factory function that can be used
+// to create new instances of timedInterval. This is useful for
+// creating multiple instances with the same processing function and
+// interval duration.
+func NewIntervalFactory[T any, R any](
+	procFunc func([]T) (R, error),
+	interval time.Duration,
+) partition.Factory[T, R] {
+	return func(
+		ctx context.Context,
+		out pipes.Pipes[R],
+		errs chan<- error,
+	) partition.Partition[T, R] {
+		return NewInterval[T, R](ctx, out, procFunc, errs, interval)
 	}
 }
